Move out the BaseFederationServlet to a separate file.
